+++ src/arena_tree.rs
@@ -34,8 +34,6 @@
 //
 
 
-#![allow(dead_code)]
-
 use std::cell::Cell;
